+++ packages/react/CalloutCard/src/calloutCard.stories.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { ComponentMeta, ComponentStory } from "@storybook/react";
 import { withA11y } from "@storybook/addon-a11y";
-import { BoxUnpackedIcon } from "@tiendanube/icons";
+import { BoxUnpackedIcon } from "@nimbus-ds/icons";
 
 import { CalloutCard } from "./CalloutCard";