+++ packages/react/FormField/src/components/FormFieldTextarea/formFieldTextarea.types.ts
@@ -1,6 +1,6 @@
 import { HTMLAttributes } from "react";
 
-import { TextareaProps } from "@nimbus-ds/textarea";
+import { TextareaProps } from "@nimbus-ds/components";
 import { FormFieldProps } from "../../formField.types";
 
 export type FormFieldTextareaProps = Omit<FormFieldProps, "children"> &