+++ packages/react/MenuButton/src/menuButton.stories.tsx
@@ -4,11 +4,8 @@ import { withA11y } from "@storybook/addon-a11y";
 // eslint-disable-next-line
 import { useArgs } from "@storybook/client-api";
 
-import { ChevronRightIcon, HomeIcon } from "@tiendanube/icons";
-import { Box } from "@nimbus-ds/box";
-import { Tag } from "@nimbus-ds/tag";
-import { Badge } from "@nimbus-ds/badge";
-import { Icon } from "@nimbus-ds/icon";
+import { ChevronRightIcon, HomeIcon } from "@nimbus-ds/icons";
+import { Box, Tag, Badge, Icon } from "@nimbus-ds/components";
 
 import { MenuButton } from "./MenuButton";
 
@@ -21,30 +18,31 @@ export default {
   argTypes: {
     label: {
       control: "text",
-      description: "Text label for the button"
+      description: "Text label for the button",
     },
     onClick: {
       control: { disable: true },
-      description: "Function executed when clicking the button"
+      description: "Function executed when clicking the button",
     },
     startIcon: {
       control: { disable: true },
-      description: "Sets an icon element on the left of the button"
+      description: "Sets an icon element on the left of the button",
     },
     children: {
       control: { disable: true },
-      description: "Inner components displayed on the right of the button component"
+      description:
+        "Inner components displayed on the right of the button component",
     },
     active: {
       control: "boolean",
       defaultValue: false,
-      description: "Sets the state of the button as active/inactive"
+      description: "Sets the state of the button as active/inactive",
     },
     disabled: {
       control: "boolean",
-      defaultValue: false
-    }
-  }
+      defaultValue: false,
+    },
+  },
 } as ComponentMeta<typeof MenuButton>;
 
 const Template: ComponentStory<typeof MenuButton> = (args) => {
@@ -60,7 +58,7 @@ const AccordionTemplate: ComponentStory<typeof MenuButton> = (args) => {
 
   return (
     <Box
-      backgroundColor={active ? 'primary.surface' : 'neutral.background'}
+      backgroundColor={active ? "primary.surface" : "neutral.background"}
       borderRadius=".5rem"
     >
       <MenuButton
@@ -93,32 +91,32 @@ const AccordionTemplate: ComponentStory<typeof MenuButton> = (args) => {
 export const sampleAppMenuButton = Template.bind({});
 sampleAppMenuButton.args = {
   label: "Menu button",
-  startIcon: HomeIcon
+  startIcon: HomeIcon,
 };
 
 export const accordionMenuButton = AccordionTemplate.bind({});
 accordionMenuButton.args = {
   label: "Menu accordion",
-  startIcon: HomeIcon
+  startIcon: HomeIcon,
 };
 
 export const menuButtonWithTag = Template.bind({});
 menuButtonWithTag.args = {
   label: "Button with tag children",
   startIcon: HomeIcon,
-  children: <Tag appearance="primary">Children tag</Tag>
+  children: <Tag appearance="primary">Children tag</Tag>,
 };
 
 export const menuButtonWithBadge = Template.bind({});
 menuButtonWithBadge.args = {
   label: "Button with badge children",
   startIcon: HomeIcon,
-  children: <Badge count="9999" appearance="primary" />
+  children: <Badge count="9999" appearance="primary" />,
 };
 
 export const menuButtonWithRightIcon = Template.bind({});
 menuButtonWithRightIcon.args = {
   label: "Button with right icon children",
   startIcon: HomeIcon,
-  children: <Icon source={<ChevronRightIcon />} />
+  children: <Icon source={<ChevronRightIcon />} />,
 };