+++ packages/react/Page/src/components/PageBody/PageBody.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { Box } from "@nimbus-ds/box";
+import { Box } from "@nimbus-ds/components";
 
 import { PageBodyProps } from "./pageBody.types";