+++ packages/react/FormField/CHANGELOG.md
@@ -2,6 +2,26 @@
 
 The FormField component allows the user to enter or edit information in text format. Additionally, it provides a label, and a conditional help text that can be used to provide additional context on a validation scenario.
 
+## 2023-02-17 `1.3.0`
+
+### 🎉 New features
+
+- Removed external dependency from `@tiendanube/icons` package to now use internal `@nimbus-ds/icons` package. ([#15](https://github.com/TiendaNube/nimbus-design-system/pull/#15) by [@juniorconquista](https://github.com/juniorconquista))
+
+### 📚 3rd party library updates
+
+- Added `@nimbus-ds/components@2.8.0`. ([#18](https://github.com/TiendaNube/nimbus-patterns/pull/18) by [@juniorconquista](https://github.com/juniorconquista))
+- Added `@nimbus-ds/styles@6.11.0`. ([#18](https://github.com/TiendaNube/nimbus-patterns/pull/18) by [@juniorconquista](https://github.com/juniorconquista))
+- Added `@nimbus-ds/icons@1.0.0`. ([#18](https://github.com/TiendaNube/nimbus-patterns/pull/18) by [@juniorconquista](https://github.com/juniorconquista))
+- Removed `@nimbus-ds/input@2.0.0`. ([#18](https://github.com/TiendaNube/nimbus-patterns/pull/18) by [@juniorconquista](https://github.com/juniorconquista))
+- Removed `@nimbus-ds/icon@2.0.0`. ([#18](https://github.com/TiendaNube/nimbus-patterns/pull/18) by [@juniorconquista](https://github.com/juniorconquista))
+- Removed `@nimbus-ds/text@5.0.0`. ([#18](https://github.com/TiendaNube/nimbus-patterns/pull/18) by [@juniorconquista](https://github.com/juniorconquista))
+- Removed `@nimbus-ds/box@2.4.0`. ([#18](https://github.com/TiendaNube/nimbus-patterns/pull/18) by [@juniorconquista](https://github.com/juniorconquista))
+- Removed `@nimbus-ds/textarea@2.0.0`. ([#18](https://github.com/TiendaNube/nimbus-patterns/pull/18) by [@juniorconquista](https://github.com/juniorconquista))
+- Removed `@nimbus-ds/label@2.0.0`. ([#18](https://github.com/TiendaNube/nimbus-patterns/pull/18) by [@juniorconquista](https://github.com/juniorconquista))
+- Removed `@nimbus-ds/select@2.0.0`. ([#18](https://github.com/TiendaNube/nimbus-patterns/pull/18) by [@juniorconquista](https://github.com/juniorconquista))
+- Removed `@tiendanube/icons@0.3.1`. ([#18](https://github.com/TiendaNube/nimbus-design-system/pull/#18) by [@juniorconquista](https://github.com/juniorconquista))
+
 ## 2023-02-10 `1.2.0`
 
 - Removed `@nimbus-ds/stack`. ([#14](https://github.com/TiendaNube/nimbus-patterns/pull/14) by [@juanchigallego](https://github.com/juanchigallego))